Fix non-deterministic behavior in MergeGepUse()
When entries are added to the worklist in MergeGepUse(), the order
is reversed to make the ordering equivalent to a recursive
traversal.
The end of the entries being reversed was off by 1, resulting in
an uninitialized or previously freed entry being shuffled into the
list. This resulted in non-determistic behavior - sometimes
producing the expected result (with extra unnecessary iterations
of the worklist loop), and sometimes resulting in a crash.